+++ libraries/muxer/src/main/java/androidx/media3/muxer/FragmentedMp4Muxer.java
@@ -22,12 +22,14 @@ import android.media.MediaCodec.BufferInfo;
 import android.util.SparseArray;
 import androidx.media3.common.Format;
 import androidx.media3.common.Metadata;
+import androidx.media3.common.MimeTypes;
 import androidx.media3.common.util.UnstableApi;
 import androidx.media3.container.MdtaMetadataEntry;
 import androidx.media3.container.Mp4LocationData;
 import androidx.media3.container.Mp4OrientationData;
 import androidx.media3.container.Mp4TimestampData;
 import androidx.media3.container.XmpData;
+import com.google.common.collect.ImmutableList;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.io.IOException;
 import java.io.OutputStream;
@@ -81,7 +83,7 @@ import java.nio.ByteBuffer;
  * </ul>
  */
 @UnstableApi
-public final class FragmentedMp4Muxer implements Muxer {
+public final class FragmentedMp4Muxer implements AutoCloseable {
   /** The default fragment duration. */
   public static final long DEFAULT_FRAGMENT_DURATION_MS = 2_000;
 
@@ -140,6 +142,24 @@ public final class FragmentedMp4Muxer implements Muxer {
     }
   }
 
+  /** A list of supported video {@linkplain MimeTypes sample MIME types}. */
+  public static final ImmutableList<String> SUPPORTED_VIDEO_SAMPLE_MIME_TYPES =
+      ImmutableList.of(
+          MimeTypes.VIDEO_AV1,
+          MimeTypes.VIDEO_H263,
+          MimeTypes.VIDEO_H264,
+          MimeTypes.VIDEO_H265,
+          MimeTypes.VIDEO_MP4V);
+
+  /** A list of supported audio {@linkplain MimeTypes sample MIME types}. */
+  public static final ImmutableList<String> SUPPORTED_AUDIO_SAMPLE_MIME_TYPES =
+      ImmutableList.of(
+          MimeTypes.AUDIO_AAC,
+          MimeTypes.AUDIO_AMR_NB,
+          MimeTypes.AUDIO_AMR_WB,
+          MimeTypes.AUDIO_OPUS,
+          MimeTypes.AUDIO_VORBIS);
+
   private final FragmentedMp4Writer fragmentedMp4Writer;
   private final MetadataCollector metadataCollector;
   private final SparseArray<Track> trackIdToTrack;
@@ -158,7 +178,14 @@ public final class FragmentedMp4Muxer implements Muxer {
     trackIdToTrack = new SparseArray<>();
   }
 
-  @Override
+  /**
+   * Adds a track of the given media format.
+   *
+   * <p>All tracks must be added before {@linkplain #writeSampleData writing any samples}.
+   *
+   * @param format The {@link Format} of the track.
+   * @return A track id for this track, which should be passed to {@link #writeSampleData}.
+   */
   public int addTrack(Format format) {
     Track track = fragmentedMp4Writer.addTrack(/* sortKey= */ 1, format);
     trackIdToTrack.append(track.id, track);
@@ -166,7 +193,7 @@ public final class FragmentedMp4Muxer implements Muxer {
   }
 
   /**
-   * {@inheritDoc}
+   * Writes encoded sample data.
    *
    * <p>Samples are written to the disk in batches. If {@link
    * Builder#setSampleCopyingEnabled(boolean) sample copying} is disabled, the {@code byteBuffer}
@@ -182,7 +209,6 @@ public final class FragmentedMp4Muxer implements Muxer {
    * @param bufferInfo The {@link BufferInfo} related to this sample.
    * @throws MuxerException If there is any error while writing data to the disk.
    */
-  @Override
   public void writeSampleData(int trackId, ByteBuffer byteBuffer, BufferInfo bufferInfo)
       throws MuxerException {
     try {
@@ -198,7 +224,7 @@ public final class FragmentedMp4Muxer implements Muxer {
   }
 
   /**
-   * {@inheritDoc}
+   * Adds {@linkplain Metadata.Entry metadata} about the output file.
    *
    * <p>List of supported {@linkplain Metadata.Entry metadata entries}:
    *
@@ -216,12 +242,18 @@ public final class FragmentedMp4Muxer implements Muxer {
    *     IllegalArgumentException} is thrown if the {@linkplain Metadata.Entry metadata} is not
    *     supported.
    */
-  @Override
   public void addMetadataEntry(Metadata.Entry metadataEntry) {
     checkArgument(MuxerUtil.isMetadataSupported(metadataEntry), "Unsupported metadata");
     metadataCollector.addMetadata(metadataEntry);
   }
 
+  /**
+   * Closes the file.
+   *
+   * <p>The muxer cannot be used anymore once this method returns.
+   *
+   * @throws MuxerException If the muxer fails to finish writing the output.
+   */
   @Override
   public void close() throws MuxerException {
     try {