+++ libraries/transformer/src/androidTest/java/androidx/media3/transformer/FrameExtractorTest.java
@@ -21,7 +21,9 @@ import static androidx.media3.exoplayer.SeekParameters.CLOSEST_SYNC;
 import static androidx.media3.test.utils.BitmapPixelTestUtil.maybeSaveTestBitmap;
 import static androidx.media3.test.utils.BitmapPixelTestUtil.readBitmap;
 import static androidx.media3.test.utils.TestUtil.assertBitmapsAreSimilar;
+import static androidx.media3.transformer.AndroidTestUtil.MP4_ASSET;
 import static androidx.media3.transformer.AndroidTestUtil.MP4_TRIM_OPTIMIZATION_270;
+import static androidx.media3.transformer.AndroidTestUtil.assumeFormatsSupported;
 import static com.google.common.truth.Truth.assertThat;
 import static com.google.common.util.concurrent.MoreExecutors.directExecutor;
 import static java.util.concurrent.TimeUnit.SECONDS;
@@ -66,7 +68,7 @@ public class FrameExtractorTest {
       "test-generated-goldens/FrameExtractorTest/";
   private static final String FILE_PATH =
       "asset:///media/mp4/sample_with_increasing_timestamps_360p.mp4";
-  private static final long TIMEOUT_SECONDS = 10;
+  private static final long TIMEOUT_SECONDS = 20;
   // TODO: b/350498258 - Due to bugs in hardware decoders, we can only assert for low PSNR values.
   // Move to using software decoders in pixel tests, and increase PSNR threshold.
   private static final float PSNR_THRESHOLD = 25f;
@@ -472,5 +474,27 @@ public class FrameExtractorTest {
     assertBitmapsAreSimilar(expectedBitmapFirstItem, actualBitmapFirstItem, PSNR_THRESHOLD);
     assertThat(frameSecondItem.presentationTimeMs).isEqualTo(8_531);
     assertBitmapsAreSimilar(expectedBitmapSecondItem, actualBitmapSecondItem, PSNR_THRESHOLD);
+  }
+
+  @Test
+  public void extractFrame_oneFrame_decodesReferenceFramesOnly() throws Exception {
+    assumeFormatsSupported(
+        context, testId, /* inputFormat= */ MP4_ASSET.videoFormat, /* outputFormat= */ null);
+    frameExtractor =
+        new ExperimentalFrameExtractor(
+            context, new ExperimentalFrameExtractor.Configuration.Builder().build());
+    frameExtractor.setMediaItem(
+        MediaItem.fromUri(MP4_ASSET.uri), /* effects= */ ImmutableList.of());
+
+    ListenableFuture<Frame> frameFuture = frameExtractor.getFrame(/* positionMs= */ 967);
+    Frame frame = frameFuture.get(TIMEOUT_SECONDS, SECONDS);
+
+    assertThat(frame.presentationTimeMs).isEqualTo(967);
+    assertThat(
+            frameExtractor
+                .getDecoderCounters()
+                .get(TIMEOUT_SECONDS, SECONDS)
+                .skippedInputBufferCount)
+        .isEqualTo(13);
   }
 }